ARM: IGEP0033: rename config file to am335x_igep0033.h
The IGEP COM AQUILA is an AM3352/AM3354 Texas Instruments
processor based industrial processor module.

It seems to be a naming convention for the configuration
files for boards using the same SoC family. This makes
easier to do changes that affect different boards based
on the same SoC.

So is better to rename IGEP COM AQUILA board config to
use this naming convention.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>