OMAP3: igep00x0: rename config file to omap3_igep00x0.h
The IGEP boards use an OMAP3535/OMAP3735 Texas Instruments
processor.

It seems to be a naming convention for the configuration
files for boards using the same SoC family. This makes
easier to do changes that affect different boards based
on the same SoC.

So, is better to rename IGEP board config to use this
naming convention.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>